requirements.txt
flask==3.0.0
flask-cors==4.0.0
openpyxl==3.1.2
google-generativeai==0.8.3
cryptography==42.0.0
pyinstaller==6.3.0
pillow==10.2.0
pyparsing==3.1.2
psutil==5.9.8
